*Added TextEditing functions as promised long ago ;) *OnScreenKeyboard improved. Now you can point to a letter and edit at that point. Also moving right left with the blinking editing line is possible with the D-PAD.
NOTE: Right now the textediting is editing one line at a time. For that click on the line you want to edit inside the TextEditor and than press the '+' button. On the OnScreenKeyboard you will have the chance to edit the line. Also to remove a linebreak remove the last char in the line (a space) and to add a new linebreak click on the Linebreak-Button on the OnScreenKeyboard.
Right now there are no text marking functions. These will come in future. This editor is still good for quick editing parts of some texts.