New Feature: RAR Extract Support with password support
Note: Using now the free rarlabs source (damn that crap is complicated). This supports RAR Formats 1.5, 2.0 and 2.9+. Right not splitted archives are not supported. This feature will follow in the future. If the archive requires a password to extract a file you will be prompted for the password. You can find the source for the lib in the branches and compile it yourself. Canceling an extraction works only between two files right now. This will also be changed later.

*Fixed bug with some ZIP archives opening empty
*Reverted usb2 source. The new one caused a few problems.
*Updated Language files